Center the stop dialog always when a stop is clicked
Call toggleTrips instead of manually toggling. Now clicking on a stop centers the marker (thanks to @paulserraino's centerMarker chinga) and scrolls the list.

Also fix bug in gulpfile. `clean` was being called every time a watch event happened. This was because `clean` was in the deps list for `build-html` and such.