Don't filter arrival times by direction (or route)
Previously, we filtered the runs (arrival times) from nextbus by route and direction.
This causes problems with all the routes because the direction IDs are switched (0 is N
for MetroRapid routes, but 1 is N for the others). Instead, just request runs for a stop
and route from nextbus and don't filter at all. We shouldn't see any runs on both
directions for a stop and route.