Find nearby stops with geohashes
This is not intended to be merged into Instabus as-is, just a proof of concept to find nearby stops quickly without extra requests.

A Geohash is a way to represent a lat, long coordinate pair as a string in base 32 (to understand how the encoding works, see: http://www.bigdatamodeling.org/2013/01/intuitive-geohash.html). Basically, two geohashes are close to each other (except for some edge cases) if they share a common prefix.

The Python data script precomputes geohashes truncated to a chosen length (7, which specifies a cell on the map 153m across on each side) for every stop, route combination. The geohashes are stored in JSON object indexed by the geohash.

Later in the browser, we load the stop geohashes when the location is found, and look for nearby routes by the current location's geohash, as well as the neighbors of the current location (enumerated based on the geohash algorithm).